Experimental: linear output
Experimental feature, might disappear again.

Linear output just walks the png pixels row by row, starting in the top
left corner. Piggybacks on hilbert code, so still only square images.

23:58 < kdave> heh, hilbert curve :)
23:59 < darkling> kdave: You can blame me for that idea. :)
23:59 < Knorrie> heh
23:59 < kdave> you! :)
23:59 < kdave> I should have figured
[...]
02:00 < kdave> I still can't get used to the hilbert type
02:01 < Knorrie> what do you want? linear?
02:01 < darkling> \o/ darkling: causing confusion since 1973.
02:01 < Knorrie> just plug in another position generator
02:01 < Knorrie> and you have another png

The resulting image gives a less well idea of locality, because there's
a big CRLF back to the left after filling a row.

02:06  * Knorrie changing it to write linear png
02:08 < kdave> a nice square box per bg, last row can be partial

If the horizontal lines in the image are too thin, it's useful to lower
the --order a bit (and resize it back up with --size if needed), so you
see a bit 'bigger boxes'.

But... this can only give nice results for smaller filesystems. If it
gets bigger, and each block group consists of less pixels, if will again
start to show contiguous used parts of disk space in a flat long line,
instead of grouping it together like the Hilbert curve does.