Automatically choose order
For smaller filesystems, it makes no sense to use order 10. It's better
to use a lower one and scale the result up a bit.

So, instead of defaulting order to 10, default size to 10 (1024x1024),
and determine order so that a single pixel represents close to 32MiB of
data. For really large filesystems, limit the auto chosen order at 10.