Show extents inside a block group
After reorganizing the code it's pretty easy to add functionality to
zoom in further and display the contents of a single block group.

We can just iterate over all the btree keys in a block group and fill in
the dots...

An EXTENT_ITEM has the length in the offset field of the key, and for
METADATA_ITEM, it's always the same, since all metadata pages are
equally sized.

Example: ./heatmap.py -vvv --blockgroup 183630823424 -o foo.png /