Allow directory as target for -o
An autogenerated filename that makes sense is nice, but I probably run
this from a cron job, and I don't want to put png images in random
locations which are cwd at the time the script runs.

So, abuse the -o option to also allow to point to a directory, in which
we put a pngfile with autogenerated name.

-# ./heatmap.py -o bla /
scope filesystem curve hilbert order 7 size 10 pngfile bla

-# rm bla
-# mkdir bla
-# ./heatmap.py -o bla /
scope filesystem curve hilbert order 7 size 10
 pngfile bla/fsid_ed10a358-c846-4e76-a071-3821d423a99d_at_1479252725.png