Don't hardcode minimal brightness for dev_extents
To be able to visually distinguish an empty block group and unallocated
space, the minimal brightness value of allocated space starts at 16
instead of 0.

But, if a user wants the difference to be bigger, for example to debug
chunk allocation behaviour, make it possible to increase that value.

Setting it to 255 will cause any allocated space to show up as bright
white, regardless of usage.

This option is only meant to be used for dev_extent level images and is
only accessible when scripting against heatmap.py:

Example:

    import btrfs
    import heatmap

    fs = btrfs.FileSystem('/')
    grid = heatmap.walk_dev_extents(fs, min_brightness=255)
    grid.write_png('all-or-nothing.png')