Extent level colors!
The location where all metadata blocks are stored are tracked in the
extent tree. A single metadata block (sized as 'nodesize' of the fs,
usually 16kB), is part of one and only one btree.

When walking the extents, we need to figure out if an extent contains
data or metadata, and if metadata, which kind of.

Using a few little tricks this can be done quite effectively.

I tried a few different palettes, and ended up with pastel colors,
similar to my vim color scheme.

This is where I got the color codes from:
https://www.akulakreative.com/blog/2012/07/the-unofficial-paper-source-color-guide