Redo and integrate Hilbert Curve
I didn't really like the hilbert curve generator, because it was a bit
too obscure, and using too many lines of code.

So, I started drawing the first few generations on paper, looking at
what happened, and then made up my own algorithm, which is based on how
the steps are replaced with multiple sub steps when moving to a higher
order.

An order 1 hilbert curve is Up, Right, Down.

    y|
     |
    0| *--->*
     | ^    |              (URD)R
     | |    v
    1| *    *
     +------------
       0    1    x

Actually, there's a fourth step, a Right, which would be the
intermediate step to do before continuing. But, since we're already
finished, we don't draw it.

Order 2 looks like this:

    y|
     |
    0| *--->*    *--->*
     | ^    |    ^    |    (RUL)U
     | |    v    |    v    (URD)R
    1| *    *--->*    *    (URD)D
     | ^              |    (LDR)R
     | |              v
    2| *<---*    *<---*
     |      ^    |
     |      |    v
    3| *--->*    *--->*
     +----------------------
       0    1    2    3    x

Each of the four steps of the order 1 curve has been expanded to another
four steps. U with RULU, etc. On the right you see the URDR back as last
letters of the four new combinations. They're the connecting steps
between each four sub curves.

Draw order 3, look at it, and connect the dots... You will end up with
the inception table as seen below.

Now, all that's left is a simple recursive function that either
substitutes the four steps with four new sets of steps, or just moves
the position around based on the four instructions, while emitting the
current position before each move.

Also, the whole thing is now just 1 file, which makes it easier to
symlink it somewhere when scripting it.

And, walking the curve is also about 25% faster than the previous code.