Show chunks in virtual address space order
This adds a sorting option to switch from sorting by physical address to
sorting by virtual address space.

What it actually does is walking the chunk tree instead of device
extents. The total size of the picture is still equal to the total size
of all attached devices, so when a chunk has multiple stripes, we just
multiply the length of the chunk by the amount of stripes.

By the way:

Extent level pictures are always sorted on virtual address, because
that's the only information we have about them. For single and raid1
profile, there is no difference, but for raid0/5/6/10 etc. there is.