Support image output to stdout
When writing the png output directly to stdout, it can be piped into a
program that can e.g. display a png in the terminal window.

The recommended program to use for this is catimg, since it uses ANSI
output instead of using ASCII characters to try drawing shapes.

For example:

    -# ./heatmap.py --size 7 / -o - | catimg -

or, of course:

    -# while :; do ./heatmap.py --size 7 / -o - | catimg -; \
       sleep 5; done

About the implementation:

Since writing the png requires going back and updating an earlier part
of the file with the length of the IDAT section, we can't just write all
output to stdout. It has to be buffered somewhere. By creating a fake
file object that quacks like a normal file, with the methods we use,
like write, seek and tell, the actual png writer code does not have to
be changed.

Thanks microraptor for suggesting this feature. This is a really nice
addition, and very useful for quickly looking at things on a remote
system.

Fixes: https://github.com/knorrie/btrfs-heatmap/issues/7